santa/requirements.txt
numpy
pandas
matplotlib
tensorflow-cpu
keras
scikit-learn